Network.AWS.KMS

Description

Amazon Key Management Service (KMS) is a managed service that makes it easy for you to create and control the encryption keys used to encrypt your data, and uses Hardware Security Modules (HSMs) to protect the security of your keys. Amazon Key Management Service is integrated with other Amazon services including Amazon EBS, Amazon S3, and Amazon Redshift. Amazon Key Management Service is also integrated with Amazon CloudTrail to provide you with logs of all key usage to help meet your regulatory and compliance needs.

Documentation

module Network.AWS.KMS.CreateAlias

module Network.AWS.KMS.CreateGrant

module Network.AWS.KMS.CreateKey

module Network.AWS.KMS.Decrypt

module Network.AWS.KMS.DeleteAlias

module Network.AWS.KMS.DescribeKey

module Network.AWS.KMS.DisableKey

module Network.AWS.KMS.DisableKeyRotation

module Network.AWS.KMS.EnableKey

module Network.AWS.KMS.EnableKeyRotation

module Network.AWS.KMS.Encrypt

module Network.AWS.KMS.GenerateDataKey

module Network.AWS.KMS.GenerateDataKeyWithoutPlaintext

module Network.AWS.KMS.GenerateRandom

module Network.AWS.KMS.GetKeyPolicy

module Network.AWS.KMS.GetKeyRotationStatus

module Network.AWS.KMS.ListAliases

module Network.AWS.KMS.ListGrants

module Network.AWS.KMS.ListKeyPolicies

module Network.AWS.KMS.ListKeys

module Network.AWS.KMS.PutKeyPolicy

module Network.AWS.KMS.ReEncrypt

module Network.AWS.KMS.RetireGrant

module Network.AWS.KMS.RevokeGrant

module Network.AWS.KMS.Types

module Network.AWS.KMS.UpdateKeyDescription