Network.AWS.KMS.CancelKeyDeletion

Description

Cancels the deletion of a customer master key (CMK). When this operation is successful, the CMK is set to the Disabled state. To enable a CMK, use EnableKey .

For more information about scheduling and canceling deletion of a CMK, see Deleting Customer Master Keys in the AWS Key Management Service Developer Guide .

Creating a Request

cancelKeyDeletion

Arguments

:: Text
ckdKeyId

-> CancelKeyDeletion

Creates a value of CancelKeyDeletion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ckdKeyId - The unique identifier for the customer master key (CMK) for which to cancel deletion. To specify this value, use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples: * Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab * Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab To obtain the unique key ID and key ARN for a given CMK, use ListKeys or DescribeKey .

data CancelKeyDeletion

See: cancelKeyDeletion smart constructor.

Request Lenses

ckdKeyId :: Lens' CancelKeyDeletion Text

The unique identifier for the customer master key (CMK) for which to cancel deletion. To specify this value, use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples: * Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab * Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab To obtain the unique key ID and key ARN for a given CMK, use ListKeys or DescribeKey .

Destructuring the Response

cancelKeyDeletionResponse

Arguments

:: Int
ckdrsResponseStatus

-> CancelKeyDeletionResponse

Creates a value of CancelKeyDeletionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ckdrsKeyId - The unique identifier of the master key for which deletion is canceled.

ckdrsResponseStatus - -- | The response status code.

data CancelKeyDeletionResponse

See: cancelKeyDeletionResponse smart constructor.

Response Lenses

ckdrsKeyId :: Lens' CancelKeyDeletionResponse (Maybe Text)

The unique identifier of the master key for which deletion is canceled.

ckdrsResponseStatus :: Lens' CancelKeyDeletionResponse Int

- | The response status code.