Network.AWS.KMS.CreateAlias

Description

Creates a display name for a customer master key. An alias can be used to identify a key and should be unique. The console enforces a one-to-one mapping between the alias and a key. An alias name can contain only alphanumeric characters, forward slashes (/), underscores (_), and dashes (-). An alias must start with the word "alias" followed by a forward slash (alias/). An alias that begins with "aws" after the forward slash (alias/aws...) is reserved by Amazon Web Services (AWS).

http://docs.aws.amazon.com/kms/latest/APIReference/API_CreateAlias.html

Request

data CreateAlias

Request constructor

createAlias

Arguments

:: Text
caAliasName

-> Text
caTargetKeyId

-> CreateAlias

CreateAlias constructor.

The fields accessible through corresponding lenses are:

caAliasName :: Text

caTargetKeyId :: Text

Request lenses

caAliasName :: Lens' CreateAlias Text

String that contains the display name. Aliases that begin with AWS are reserved.

caTargetKeyId :: Lens' CreateAlias Text

An identifier of the key for which you are creating the alias. This value cannot be another alias but can be a globally unique identifier or a fully specified ARN to a key. Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012 Globally Unique Key ID Example - 12345678-1234-1234-123456789012

Response

data CreateAliasResponse

Response constructor

createAliasResponse :: CreateAliasResponse

CreateAliasResponse constructor.