Network.AWS.KMS.CreateKey

Description

Creates a customer master key. Customer master keys can be used to encrypt small amounts of data (less than 4K) directly, but they are most commonly used to encrypt or envelope data keys that are then used to encrypt customer data. For more information about data keys, see GenerateDataKey and GenerateDataKeyWithoutPlaintext.

http://docs.aws.amazon.com/kms/latest/APIReference/API_CreateKey.html

Request

data CreateKey

Request constructor

createKey :: CreateKey

CreateKey constructor.

The fields accessible through corresponding lenses are:

ckDescription :: Maybe Text

ckKeyUsage :: Maybe KeyUsageType

ckPolicy :: Maybe Text

Request lenses

ckDescription :: Lens' CreateKey (Maybe Text)

Description of the key. We recommend that you choose a description that helps your customer decide whether the key is appropriate for a task.

ckKeyUsage :: Lens' CreateKey (Maybe KeyUsageType)

Specifies the intended use of the key. Currently this defaults to ENCRYPT/DECRYPT, and only symmetric encryption and decryption are supported.

ckPolicy :: Lens' CreateKey (Maybe Text)

Policy to be attached to the key. This is required and delegates back to the account. The key is the root of trust.

Response

data CreateKeyResponse

Response constructor

createKeyResponse :: CreateKeyResponse

CreateKeyResponse constructor.

The fields accessible through corresponding lenses are:

ckrKeyMetadata :: Maybe KeyMetadata

Response lenses

ckrKeyMetadata :: Lens' CreateKeyResponse (Maybe KeyMetadata)

Metadata associated with the key.