Network.AWS.KMS.Decrypt

Description

Decrypts ciphertext. Ciphertext is plaintext that has been previously encrypted by using the Encrypt function.

http://docs.aws.amazon.com/kms/latest/APIReference/API_Decrypt.html

Request

data Decrypt

Request constructor

decrypt

Arguments

:: Base64
dCiphertextBlob

-> Decrypt

Decrypt constructor.

The fields accessible through corresponding lenses are:

dCiphertextBlob :: Base64

dEncryptionContext :: HashMap Text Text

dGrantTokens :: [Text]

Request lenses

dCiphertextBlob :: Lens' Decrypt Base64

Ciphertext including metadata.

dEncryptionContext :: Lens' Decrypt (HashMap Text Text)

The encryption context. If this was specified in the Encrypt function, it must be specified here or the decryption operation will fail. For more information, see Encryption Context.

dGrantTokens :: Lens' Decrypt [Text]

A list of grant tokens that represent grants which can be used to provide long term permissions to perform decryption.

Response

data DecryptResponse

Response constructor

decryptResponse :: DecryptResponse

DecryptResponse constructor.

The fields accessible through corresponding lenses are:

drKeyId :: Maybe Text

drPlaintext :: Maybe Base64

Response lenses

drKeyId :: Lens' DecryptResponse (Maybe Text)

Unique identifier created by the system for the key. This value is always returned as long as no errors are encountered during the operation.

drPlaintext :: Lens' DecryptResponse (Maybe Base64)

Decrypted plaintext data. This value may not be returned if the customer master key is not available or if you didn't have permission to use it.