Network.AWS.KMS.DeleteImportedKeyMaterial

Description

Deletes key material that you previously imported and makes the specified customer master key (CMK) unusable. For more information about importing key material into AWS KMS, see Importing Key Material in the AWS Key Management Service Developer Guide .

When the specified CMK is in the PendingDeletion state, this operation does not change the CMK's state. Otherwise, it changes the CMK's state to PendingImport .

After you delete key material, you can use ImportKeyMaterial to reimport the same key material into the CMK.

Creating a Request

deleteImportedKeyMaterial

Arguments

:: Text
dikmKeyId

-> DeleteImportedKeyMaterial

Creates a value of DeleteImportedKeyMaterial with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dikmKeyId - The identifier of the CMK whose key material to delete. The CMK's Origin must be EXTERNAL . A valid identifier is the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples: * Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab * Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

data DeleteImportedKeyMaterial

See: deleteImportedKeyMaterial smart constructor.

Request Lenses

dikmKeyId :: Lens' DeleteImportedKeyMaterial Text

The identifier of the CMK whose key material to delete. The CMK's Origin must be EXTERNAL . A valid identifier is the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples: * Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab * Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Destructuring the Response

deleteImportedKeyMaterialResponse :: DeleteImportedKeyMaterialResponse

Creates a value of DeleteImportedKeyMaterialResponse with the minimum fields required to make a request.

data DeleteImportedKeyMaterialResponse

See: deleteImportedKeyMaterialResponse smart constructor.