Network.AWS.KMS.DescribeKey

Description

Provides detailed information about the specified customer master key.

http://docs.aws.amazon.com/kms/latest/APIReference/API_DescribeKey.html

Request

data DescribeKey

Request constructor

describeKey

Arguments

:: Text
dk1KeyId

-> DescribeKey

DescribeKey constructor.

The fields accessible through corresponding lenses are:

dk1KeyId :: Text

Request lenses

dk1KeyId :: Lens' DescribeKey Text

A unique identifier for the customer master key. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/". Key ARN Example - arn:aws:kms:us-east-1:123456789012:key12345678-1234-1234-1234-123456789012 Alias ARN Example - arn:aws:kms:us-east-1:123456789012:alias/MyAliasName Globally Unique Key ID Example - 12345678-1234-1234-123456789012 Alias Name Example - alias/MyAliasName

Response

data DescribeKeyResponse

Response constructor

describeKeyResponse :: DescribeKeyResponse

DescribeKeyResponse constructor.

The fields accessible through corresponding lenses are:

dkrKeyMetadata :: Maybe KeyMetadata

Response lenses

dkrKeyMetadata :: Lens' DescribeKeyResponse (Maybe KeyMetadata)

Metadata associated with the key.