Network.AWS.KMS.EnableKey

Description

Marks a key as enabled, thereby permitting its use. You can have up to 25 enabled keys at one time.

http://docs.aws.amazon.com/kms/latest/APIReference/API_EnableKey.html

Request

data EnableKey

Request constructor

enableKey

Arguments

:: Text
ekKeyId

-> EnableKey

EnableKey constructor.

The fields accessible through corresponding lenses are:

ekKeyId :: Text

Request lenses

ekKeyId :: Lens' EnableKey Text

Unique identifier of the customer master key to be enabled. This can be an ARN, an alias, or a globally unique identifier.

Response

data EnableKeyResponse

Response constructor

enableKeyResponse :: EnableKeyResponse

EnableKeyResponse constructor.