Network.AWS.KMS.EnableKeyRotation

Description

Enables rotation of the specified customer master key.

Creating a Request

enableKeyRotation

Arguments

:: Text
ekrKeyId

-> EnableKeyRotation

Creates a value of EnableKeyRotation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ekrKeyId - A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key. * Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012 * Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

data EnableKeyRotation

See: enableKeyRotation smart constructor.

Request Lenses

ekrKeyId :: Lens' EnableKeyRotation Text

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key. * Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012 * Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Destructuring the Response

enableKeyRotationResponse :: EnableKeyRotationResponse

Creates a value of EnableKeyRotationResponse with the minimum fields required to make a request.

data EnableKeyRotationResponse

See: enableKeyRotationResponse smart constructor.