Network.AWS.KMS.Encrypt

Description

Encrypts plaintext into ciphertext by using a customer master key.

http://docs.aws.amazon.com/kms/latest/APIReference/API_Encrypt.html

Request

data Encrypt

Request constructor

encrypt

Arguments

:: Text
eKeyId

-> Base64
ePlaintext

-> Encrypt

Encrypt constructor.

The fields accessible through corresponding lenses are:

eEncryptionContext :: HashMap Text Text

eGrantTokens :: [Text]

eKeyId :: Text

ePlaintext :: Base64

Request lenses

eEncryptionContext :: Lens' Encrypt (HashMap Text Text)

Name:value pair that specifies the encryption context to be used for authenticated encryption. For more information, see <http://docs.aws.amazon.com/kms/latest/developerguide/crypto_authen.html Authenticated Encryption>.

eGrantTokens :: Lens' Encrypt [Text]

A list of grant tokens that represent grants which can be used to provide long term permissions to perform encryption.

eKeyId :: Lens' Encrypt Text

Unique identifier of the customer master. This can be an ARN, an alias, or the Key ID.

ePlaintext :: Lens' Encrypt Base64

Data to be encrypted.

Response

data EncryptResponse

Response constructor

encryptResponse :: EncryptResponse

EncryptResponse constructor.

The fields accessible through corresponding lenses are:

erCiphertextBlob :: Maybe Base64

erKeyId :: Maybe Text

Response lenses

erCiphertextBlob :: Lens' EncryptResponse (Maybe Base64)

The encrypted plaintext.

erKeyId :: Lens' EncryptResponse (Maybe Text)

The ID of the key used during encryption.