Network.AWS.KMS.GetKeyPolicy

Description

Retrieves a policy attached to the specified key.

Creating a Request

getKeyPolicy

Arguments

:: Text
gkpKeyId

-> Text
gkpPolicyName

-> GetKeyPolicy

Creates a value of GetKeyPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gkpKeyId - A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key. * Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012 * Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

gkpPolicyName - String that contains the name of the policy. Currently, this must be "default". Policy names can be discovered by calling ListKeyPolicies .

data GetKeyPolicy

See: getKeyPolicy smart constructor.

Request Lenses

gkpKeyId :: Lens' GetKeyPolicy Text

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key. * Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012 * Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

gkpPolicyName :: Lens' GetKeyPolicy Text

String that contains the name of the policy. Currently, this must be "default". Policy names can be discovered by calling ListKeyPolicies .

Destructuring the Response

getKeyPolicyResponse

Arguments

:: Int
gkprsResponseStatus

-> GetKeyPolicyResponse

Creates a value of GetKeyPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gkprsPolicy - A policy document in JSON format.

gkprsResponseStatus - -- | The response status code.

data GetKeyPolicyResponse

See: getKeyPolicyResponse smart constructor.

Response Lenses

gkprsPolicy :: Lens' GetKeyPolicyResponse (Maybe Text)

A policy document in JSON format.

gkprsResponseStatus :: Lens' GetKeyPolicyResponse Int

- | The response status code.