Network.AWS.KMS.GetKeyRotationStatus

Description

Retrieves a Boolean value that indicates whether key rotation is enabled for the specified key.

http://docs.aws.amazon.com/kms/latest/APIReference/API_GetKeyRotationStatus.html

Request

data GetKeyRotationStatus

Request constructor

getKeyRotationStatus

Arguments

:: Text
gkrsKeyId

-> GetKeyRotationStatus

GetKeyRotationStatus constructor.

The fields accessible through corresponding lenses are:

gkrsKeyId :: Text

Request lenses

gkrsKeyId :: Lens' GetKeyRotationStatus Text

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key. Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012 Globally Unique Key ID Example - 12345678-1234-1234-123456789012

Response

data GetKeyRotationStatusResponse

Response constructor

getKeyRotationStatusResponse :: GetKeyRotationStatusResponse

GetKeyRotationStatusResponse constructor.

The fields accessible through corresponding lenses are:

gkrsrKeyRotationEnabled :: Maybe Bool

Response lenses

gkrsrKeyRotationEnabled :: Lens' GetKeyRotationStatusResponse (Maybe Bool)

A Boolean value that specifies whether key rotation is enabled.