Network.AWS.KMS.ListAliases

Description

Lists all of the key aliases in the account.

http://docs.aws.amazon.com/kms/latest/APIReference/API_ListAliases.html

Request

data ListAliases

Request constructor

listAliases :: ListAliases

ListAliases constructor.

The fields accessible through corresponding lenses are:

laLimit :: Maybe Natural

laMarker :: Maybe Text

Request lenses

laLimit :: Lens' ListAliases (Maybe Natural)

Specify this parameter when paginating results to indicate the maximum number of aliases you want in each response. If there are additional aliases beyond the maximum you specify, the Truncated response element will be set to 'true.'

laMarker :: Lens' ListAliases (Maybe Text)

Use this parameter when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the NextMarker element in the response you just received.

Response

data ListAliasesResponse

Response constructor

listAliasesResponse :: ListAliasesResponse

ListAliasesResponse constructor.

The fields accessible through corresponding lenses are:

larAliases :: [AliasListEntry]

larNextMarker :: Maybe Text

larTruncated :: Maybe Bool

Response lenses

larAliases :: Lens' ListAliasesResponse [AliasListEntry]

A list of key aliases in the user's account.

larNextMarker :: Lens' ListAliasesResponse (Maybe Text)

If Truncated is true, this value is present and contains the value to use for the Marker request parameter in a subsequent pagination request.

larTruncated :: Lens' ListAliasesResponse (Maybe Bool)

A flag that indicates whether there are more items in the list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more aliases in the list.