Network.AWS.KMS.ListGrants

Description

List the grants for a specified key.

http://docs.aws.amazon.com/kms/latest/APIReference/API_ListGrants.html

Request

data ListGrants

Request constructor

listGrants

Arguments

:: Text
lgKeyId

-> ListGrants

ListGrants constructor.

The fields accessible through corresponding lenses are:

lgKeyId :: Text

lgLimit :: Maybe Natural

lgMarker :: Maybe Text

Request lenses

lgKeyId :: Lens' ListGrants Text

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key. Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012 Globally Unique Key ID Example - 12345678-1234-1234-123456789012

lgLimit :: Lens' ListGrants (Maybe Natural)

Specify this parameter only when paginating results to indicate the maximum number of grants you want listed in the response. If there are additional grants beyond the maximum you specify, the Truncated response element will be set to 'true.'

lgMarker :: Lens' ListGrants (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the NextMarker in the response you just received.

Response

data ListGrantsResponse

Response constructor

listGrantsResponse :: ListGrantsResponse

ListGrantsResponse constructor.

The fields accessible through corresponding lenses are:

lgrGrants :: [GrantListEntry]

lgrNextMarker :: Maybe Text

lgrTruncated :: Maybe Bool

Response lenses

lgrGrants :: Lens' ListGrantsResponse [GrantListEntry]

A list of grants.

lgrNextMarker :: Lens' ListGrantsResponse (Maybe Text)

If Truncated is true, this value is present and contains the value to use for the Marker request parameter in a subsequent pagination request.

lgrTruncated :: Lens' ListGrantsResponse (Maybe Bool)

A flag that indicates whether there are more items in the list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more grants in the list.