Network.AWS.KMS.ListKeyPolicies

Description

Retrieves a list of policies attached to a key.

http://docs.aws.amazon.com/kms/latest/APIReference/API_ListKeyPolicies.html

Request

data ListKeyPolicies

Request constructor

listKeyPolicies

Arguments

:: Text
lkpKeyId

-> ListKeyPolicies

ListKeyPolicies constructor.

The fields accessible through corresponding lenses are:

lkpKeyId :: Text

lkpLimit :: Maybe Natural

lkpMarker :: Maybe Text

Request lenses

lkpKeyId :: Lens' ListKeyPolicies Text

A unique identifier for the customer master key. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/". Key ARN Example - arn:aws:kms:us-east-1:123456789012:key12345678-1234-1234-1234-123456789012 Alias ARN Example - arn:aws:kms:us-east-1:123456789012:alias/MyAliasName Globally Unique Key ID Example - 12345678-1234-1234-123456789012 Alias Name Example - alias/MyAliasName

lkpLimit :: Lens' ListKeyPolicies (Maybe Natural)

Specify this parameter only when paginating results to indicate the maximum number of policies you want listed in the response. If there are additional policies beyond the maximum you specify, the Truncated response element will be set to 'true.'

lkpMarker :: Lens' ListKeyPolicies (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the NextMarker in the response you just received.

Response

data ListKeyPoliciesResponse

Response constructor

listKeyPoliciesResponse :: ListKeyPoliciesResponse

ListKeyPoliciesResponse constructor.

The fields accessible through corresponding lenses are:

lkprNextMarker :: Maybe Text

lkprPolicyNames :: [Text]

lkprTruncated :: Maybe Bool

Response lenses

lkprNextMarker :: Lens' ListKeyPoliciesResponse (Maybe Text)

If Truncated is true, this value is present and contains the value to use for the Marker request parameter in a subsequent pagination request.

lkprPolicyNames :: Lens' ListKeyPoliciesResponse [Text]

A list of policy names. Currently, there is only one policy and it is named Default.

lkprTruncated :: Lens' ListKeyPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more items in the list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more policies in the list.