Network.AWS.KMS.ListKeys

Description

Lists the customer master keys.

http://docs.aws.amazon.com/kms/latest/APIReference/API_ListKeys.html

Request

data ListKeys

Request constructor

listKeys :: ListKeys

ListKeys constructor.

The fields accessible through corresponding lenses are:

lkLimit :: Maybe Natural

lkMarker :: Maybe Text

Request lenses

lkLimit :: Lens' ListKeys (Maybe Natural)

Specify this parameter only when paginating results to indicate the maximum number of keys you want listed in the response. If there are additional keys beyond the maximum you specify, the Truncated response element will be set to 'true.'

lkMarker :: Lens' ListKeys (Maybe Text)

Use this parameter only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the NextMarker in the response you just received.

Response

data ListKeysResponse

Response constructor

listKeysResponse :: ListKeysResponse

ListKeysResponse constructor.

The fields accessible through corresponding lenses are:

lkrKeys :: [KeyListEntry]

lkrNextMarker :: Maybe Text

lkrTruncated :: Maybe Bool

Response lenses

lkrKeys :: Lens' ListKeysResponse [KeyListEntry]

A list of keys.

lkrNextMarker :: Lens' ListKeysResponse (Maybe Text)

If Truncated is true, this value is present and contains the value to use for the Marker request parameter in a subsequent pagination request.

lkrTruncated :: Lens' ListKeysResponse (Maybe Bool)

A flag that indicates whether there are more items in the list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more keys in the list.