Network.AWS.KMS.PutKeyPolicy

Description

Attaches a key policy to the specified customer master key (CMK).

For more information about key policies, see Key Policies in the AWS Key Management Service Developer Guide .

Creating a Request

putKeyPolicy

Arguments

:: Text
pkpKeyId

-> Text
pkpPolicyName

-> Text
pkpPolicy

-> PutKeyPolicy

Creates a value of PutKeyPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pkpBypassPolicyLockoutSafetyCheck - A flag to indicate whether to bypass the key policy lockout safety check. Important: Setting this value to true increases the likelihood that the CMK becomes unmanageable. Do not set this value to true indiscriminately. For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide . Use this parameter only when you intend to prevent the principal that is making the request from making a subsequent PutKeyPolicy request on the CMK. The default value is false.

pkpKeyId - A unique identifier for the CMK. Use the CMK's unique identifier or its Amazon Resource Name (ARN). For example: * Unique ID: 1234abcd-12ab-34cd-56ef-1234567890ab * ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

pkpPolicyName - The name of the key policy. This value must be default .

pkpPolicy - The key policy to attach to the CMK. If you do not set BypassPolicyLockoutSafetyCheck to true, the policy must meet the following criteria: * It must allow the principal that is making the PutKeyPolicy request to make a subsequent PutKeyPolicy request on the CMK. This reduces the likelihood that the CMK becomes unmanageable. For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide . * The principals that are specified in the key policy must exist and be visible to AWS KMS. When you create a new AWS principal (for example, an IAM user or role), you might need to enforce a delay before specifying the new principal in a key policy because the new principal might not immediately be visible to AWS KMS. For more information, see Changes that I make are not always immediately visible in the IAM User Guide . The policy size limit is 32 KiB (32768 bytes).

data PutKeyPolicy

See: putKeyPolicy smart constructor.

Request Lenses

pkpBypassPolicyLockoutSafetyCheck :: Lens' PutKeyPolicy (Maybe Bool)

A flag to indicate whether to bypass the key policy lockout safety check. Important: Setting this value to true increases the likelihood that the CMK becomes unmanageable. Do not set this value to true indiscriminately. For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide . Use this parameter only when you intend to prevent the principal that is making the request from making a subsequent PutKeyPolicy request on the CMK. The default value is false.

pkpKeyId :: Lens' PutKeyPolicy Text

A unique identifier for the CMK. Use the CMK's unique identifier or its Amazon Resource Name (ARN). For example: * Unique ID: 1234abcd-12ab-34cd-56ef-1234567890ab * ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

pkpPolicyName :: Lens' PutKeyPolicy Text

The name of the key policy. This value must be default .

pkpPolicy :: Lens' PutKeyPolicy Text

The key policy to attach to the CMK. If you do not set BypassPolicyLockoutSafetyCheck to true, the policy must meet the following criteria: * It must allow the principal that is making the PutKeyPolicy request to make a subsequent PutKeyPolicy request on the CMK. This reduces the likelihood that the CMK becomes unmanageable. For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide . * The principals that are specified in the key policy must exist and be visible to AWS KMS. When you create a new AWS principal (for example, an IAM user or role), you might need to enforce a delay before specifying the new principal in a key policy because the new principal might not immediately be visible to AWS KMS. For more information, see Changes that I make are not always immediately visible in the IAM User Guide . The policy size limit is 32 KiB (32768 bytes).

Destructuring the Response

putKeyPolicyResponse :: PutKeyPolicyResponse

Creates a value of PutKeyPolicyResponse with the minimum fields required to make a request.

data PutKeyPolicyResponse

See: putKeyPolicyResponse smart constructor.