Network.AWS.KMS.RevokeGrant

Description

Revokes a grant. You can revoke a grant to actively deny operations that depend on it.

http://docs.aws.amazon.com/kms/latest/APIReference/API_RevokeGrant.html

Request

data RevokeGrant

Request constructor

revokeGrant

Arguments

:: Text
rgKeyId

-> Text
rgGrantId

-> RevokeGrant

RevokeGrant constructor.

The fields accessible through corresponding lenses are:

rgGrantId :: Text

rgKeyId :: Text

Request lenses

rgGrantId :: Lens' RevokeGrant Text

Identifier of the grant to be revoked.

rgKeyId :: Lens' RevokeGrant Text

A unique identifier for the customer master key associated with the grant. This value can be a globally unique identifier or the fully specified ARN to a key. Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012 Globally Unique Key ID Example - 12345678-1234-1234-123456789012

Response

data RevokeGrantResponse

Response constructor

revokeGrantResponse :: RevokeGrantResponse

RevokeGrantResponse constructor.