Network.AWS.KMS.ScheduleKeyDeletion

Description

Schedules the deletion of a customer master key (CMK). You may provide a waiting period, specified in days, before deletion occurs. If you do not provide a waiting period, the default period of 30 days is used. When this operation is successful, the state of the CMK changes to PendingDeletion . Before the waiting period ends, you can use CancelKeyDeletion to cancel the deletion of the CMK. After the waiting period ends, AWS KMS deletes the CMK and all AWS KMS data associated with it, including all aliases that refer to it.

Important: Deleting a CMK is a destructive and potentially dangerous operation. When a CMK is deleted, all data that was encrypted under the CMK is rendered unrecoverable. To restrict the use of a CMK without deleting it, use DisableKey .

For more information about scheduling a CMK for deletion, see Deleting Customer Master Keys in the AWS Key Management Service Developer Guide .

Creating a Request

scheduleKeyDeletion

Arguments

:: Text
skdKeyId

-> ScheduleKeyDeletion

Creates a value of ScheduleKeyDeletion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

skdPendingWindowInDays - The waiting period, specified in number of days. After the waiting period ends, AWS KMS deletes the customer master key (CMK). This value is optional. If you include a value, it must be between 7 and 30, inclusive. If you do not include a value, it defaults to 30.

skdKeyId - The unique identifier for the customer master key (CMK) to delete. To specify this value, use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples: * Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab * Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab To obtain the unique key ID and key ARN for a given CMK, use ListKeys or DescribeKey .

data ScheduleKeyDeletion

See: scheduleKeyDeletion smart constructor.

Request Lenses

skdPendingWindowInDays :: Lens' ScheduleKeyDeletion (Maybe Natural)

The waiting period, specified in number of days. After the waiting period ends, AWS KMS deletes the customer master key (CMK). This value is optional. If you include a value, it must be between 7 and 30, inclusive. If you do not include a value, it defaults to 30.

skdKeyId :: Lens' ScheduleKeyDeletion Text

The unique identifier for the customer master key (CMK) to delete. To specify this value, use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples: * Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab * Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab To obtain the unique key ID and key ARN for a given CMK, use ListKeys or DescribeKey .

Destructuring the Response

scheduleKeyDeletionResponse

Arguments

:: Int
skdrsResponseStatus

-> ScheduleKeyDeletionResponse

Creates a value of ScheduleKeyDeletionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

skdrsKeyId - The unique identifier of the customer master key (CMK) for which deletion is scheduled.

skdrsDeletionDate - The date and time after which AWS KMS deletes the customer master key (CMK).

skdrsResponseStatus - -- | The response status code.

data ScheduleKeyDeletionResponse

See: scheduleKeyDeletionResponse smart constructor.

Response Lenses

skdrsKeyId :: Lens' ScheduleKeyDeletionResponse (Maybe Text)

The unique identifier of the customer master key (CMK) for which deletion is scheduled.

skdrsDeletionDate :: Lens' ScheduleKeyDeletionResponse (Maybe UTCTime)

The date and time after which AWS KMS deletes the customer master key (CMK).

skdrsResponseStatus :: Lens' ScheduleKeyDeletionResponse Int

- | The response status code.