Network.AWS.KMS.Types

Service

data KMS

Version 2014-11-01 of the Amazon Key Management Service service.

Error

data JSONError :: *

KeyUsageType

data KeyUsageType

Constructors

EncryptDecrypt
ENCRYPT_DECRYPT

KeyMetadata

data KeyMetadata

keyMetadata

Arguments

:: Text
kmKeyId

-> KeyMetadata

KeyMetadata constructor.

The fields accessible through corresponding lenses are:

kmAWSAccountId :: Maybe Text

kmArn :: Maybe Text

kmCreationDate :: Maybe UTCTime

kmDescription :: Maybe Text

kmEnabled :: Maybe Bool

kmKeyId :: Text

kmKeyUsage :: Maybe KeyUsageType

kmAWSAccountId :: Lens' KeyMetadata (Maybe Text)

Account ID number.

kmArn :: Lens' KeyMetadata (Maybe Text)

Key ARN (Amazon Resource Name).

kmCreationDate :: Lens' KeyMetadata (Maybe UTCTime)

Date the key was created.

kmDescription :: Lens' KeyMetadata (Maybe Text)

The description of the key.

kmEnabled :: Lens' KeyMetadata (Maybe Bool)

Value that specifies whether the key is enabled.

kmKeyId :: Lens' KeyMetadata Text

Unique identifier for the key.

kmKeyUsage :: Lens' KeyMetadata (Maybe KeyUsageType)

A value that specifies what operation(s) the key can perform.

DataKeySpec

data DataKeySpec

Constructors

AES128
AES_128

AES256
AES_256

GrantConstraints

data GrantConstraints

grantConstraints :: GrantConstraints

GrantConstraints constructor.

The fields accessible through corresponding lenses are:

gcEncryptionContextEquals :: HashMap Text Text

gcEncryptionContextSubset :: HashMap Text Text

gcEncryptionContextEquals :: Lens' GrantConstraints (HashMap Text Text)

The constraint contains additional key/value pairs that serve to further limit the grant.

gcEncryptionContextSubset :: Lens' GrantConstraints (HashMap Text Text)

The constraint equals the full encryption context.

AliasListEntry

data AliasListEntry

aliasListEntry :: AliasListEntry

AliasListEntry constructor.

The fields accessible through corresponding lenses are:

aleAliasArn :: Maybe Text

aleAliasName :: Maybe Text

aleTargetKeyId :: Maybe Text

aleAliasArn :: Lens' AliasListEntry (Maybe Text)

String that contains the key ARN.

aleAliasName :: Lens' AliasListEntry (Maybe Text)

String that contains the alias.

aleTargetKeyId :: Lens' AliasListEntry (Maybe Text)

String that contains the key identifier pointed to by the alias.

GrantListEntry

data GrantListEntry

grantListEntry :: GrantListEntry

GrantListEntry constructor.

The fields accessible through corresponding lenses are:

gleConstraints :: Maybe GrantConstraints

gleGrantId :: Maybe Text

gleGranteePrincipal :: Maybe Text

gleIssuingAccount :: Maybe Text

gleOperations :: [GrantOperation]

gleRetiringPrincipal :: Maybe Text

gleConstraints :: Lens' GrantListEntry (Maybe GrantConstraints)

Specifies the conditions under which the actions specified by the Operations parameter are allowed.

gleGrantId :: Lens' GrantListEntry (Maybe Text)

Unique grant identifier.

gleGranteePrincipal :: Lens' GrantListEntry (Maybe Text)

The principal that receives the grant permission.

gleIssuingAccount :: Lens' GrantListEntry (Maybe Text)

The account under which the grant was issued.

gleOperations :: Lens' GrantListEntry [GrantOperation]

List of operations permitted by the grant. This can be any combination of one or more of the following values: Decrypt Encrypt GenerateDataKey GenerateDataKeyWithoutPlaintext ReEncryptFrom ReEncryptTo CreateGrant

gleRetiringPrincipal :: Lens' GrantListEntry (Maybe Text)

The principal that can retire the account.

GrantOperation

data GrantOperation

Constructors

GOCreateGrant
CreateGrant

GODecrypt
Decrypt

GOEncrypt
Encrypt

GOGenerateDataKey
GenerateDataKey

GOGenerateDataKeyWithoutPlaintext
GenerateDataKeyWithoutPlaintext

GOReEncryptFrom
ReEncryptFrom

GOReEncryptTo
ReEncryptTo

GORetireGrant
RetireGrant

KeyListEntry

data KeyListEntry

keyListEntry :: KeyListEntry

KeyListEntry constructor.

The fields accessible through corresponding lenses are:

kleKeyArn :: Maybe Text

kleKeyId :: Maybe Text

kleKeyArn :: Lens' KeyListEntry (Maybe Text)

ARN of the key.

kleKeyId :: Lens' KeyListEntry (Maybe Text)

Unique identifier of the key.