Network.AWS.KMS.CreateGrant

Description

Adds a grant to a key to specify who can use the key and under what conditions. Grants are alternate permission mechanisms to key policies.

For more information about grants, see Grants in the AWS Key Management Service Developer Guide .

Creating a Request

createGrant

Arguments

:: Text
cgKeyId

-> Text
cgGranteePrincipal

-> CreateGrant

Creates a value of CreateGrant with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgRetiringPrincipal - The principal that is given permission to retire the grant by using RetireGrant operation. To specify the principal, use the Amazon Resource Name (ARN) of an AWS principal. Valid AWS principals include AWS accounts (root), IAM users, federated users, and assumed role users. For examples of the ARN syntax to use for specifying a principal, see AWS Identity and Access Management (IAM) in the Example ARNs section of the AWS General Reference .

cgGrantTokens - A list of grant tokens. For more information, see Grant Tokens in the AWS Key Management Service Developer Guide .

cgConstraints - A structure that you can use to allow certain operations in the grant only when the desired encryption context is present. For more information about encryption context, see Encryption Context in the AWS Key Management Service Developer Guide .

cgName - A friendly name for identifying the grant. Use this value to prevent unintended creation of duplicate grants when retrying this request. When this value is absent, all CreateGrant requests result in a new grant with a unique GrantId even if all the supplied parameters are identical. This can result in unintended duplicates when you retry the CreateGrant request. When this value is present, you can retry a CreateGrant request with identical parameters; if the grant already exists, the original GrantId is returned without creating a new grant. Note that the returned grant token is unique with every CreateGrant request, even when a duplicate GrantId is returned. All grant tokens obtained in this way can be used interchangeably.

cgOperations - A list of operations that the grant permits.

cgKeyId - The unique identifier for the customer master key (CMK) that the grant applies to. To specify this value, use the globally unique key ID or the Amazon Resource Name (ARN) of the key. Examples: * Globally unique key ID: 12345678-1234-1234-1234-123456789012 * Key ARN: arn:aws:kms:us-west-2:123456789012:key/12345678-1234-1234-1234-123456789012

cgGranteePrincipal - The principal that is given permission to perform the operations that the grant permits. To specify the principal, use the Amazon Resource Name (ARN) of an AWS principal. Valid AWS principals include AWS accounts (root), IAM users, IAM roles, federated users, and assumed role users. For examples of the ARN syntax to use for specifying a principal, see AWS Identity and Access Management (IAM) in the Example ARNs section of the AWS General Reference .

data CreateGrant

See: createGrant smart constructor.

Request Lenses

cgRetiringPrincipal :: Lens' CreateGrant (Maybe Text)

The principal that is given permission to retire the grant by using RetireGrant operation. To specify the principal, use the Amazon Resource Name (ARN) of an AWS principal. Valid AWS principals include AWS accounts (root), IAM users, federated users, and assumed role users. For examples of the ARN syntax to use for specifying a principal, see AWS Identity and Access Management (IAM) in the Example ARNs section of the AWS General Reference .

cgGrantTokens :: Lens' CreateGrant [Text]

A list of grant tokens. For more information, see Grant Tokens in the AWS Key Management Service Developer Guide .

cgConstraints :: Lens' CreateGrant (Maybe GrantConstraints)

A structure that you can use to allow certain operations in the grant only when the desired encryption context is present. For more information about encryption context, see Encryption Context in the AWS Key Management Service Developer Guide .

cgName :: Lens' CreateGrant (Maybe Text)

A friendly name for identifying the grant. Use this value to prevent unintended creation of duplicate grants when retrying this request. When this value is absent, all CreateGrant requests result in a new grant with a unique GrantId even if all the supplied parameters are identical. This can result in unintended duplicates when you retry the CreateGrant request. When this value is present, you can retry a CreateGrant request with identical parameters; if the grant already exists, the original GrantId is returned without creating a new grant. Note that the returned grant token is unique with every CreateGrant request, even when a duplicate GrantId is returned. All grant tokens obtained in this way can be used interchangeably.

cgOperations :: Lens' CreateGrant [GrantOperation]

A list of operations that the grant permits.

cgKeyId :: Lens' CreateGrant Text

The unique identifier for the customer master key (CMK) that the grant applies to. To specify this value, use the globally unique key ID or the Amazon Resource Name (ARN) of the key. Examples: * Globally unique key ID: 12345678-1234-1234-1234-123456789012 * Key ARN: arn:aws:kms:us-west-2:123456789012:key/12345678-1234-1234-1234-123456789012

cgGranteePrincipal :: Lens' CreateGrant Text

The principal that is given permission to perform the operations that the grant permits. To specify the principal, use the Amazon Resource Name (ARN) of an AWS principal. Valid AWS principals include AWS accounts (root), IAM users, IAM roles, federated users, and assumed role users. For examples of the ARN syntax to use for specifying a principal, see AWS Identity and Access Management (IAM) in the Example ARNs section of the AWS General Reference .

Destructuring the Response

createGrantResponse

Arguments

:: Int
cgrsResponseStatus

-> CreateGrantResponse

Creates a value of CreateGrantResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgrsGrantId - The unique identifier for the grant. You can use the GrantId in a subsequent RetireGrant or RevokeGrant operation.

cgrsGrantToken - The grant token. For more information, see Grant Tokens in the AWS Key Management Service Developer Guide .

cgrsResponseStatus - -- | The response status code.

data CreateGrantResponse

See: createGrantResponse smart constructor.

Response Lenses

cgrsGrantId :: Lens' CreateGrantResponse (Maybe Text)

The unique identifier for the grant. You can use the GrantId in a subsequent RetireGrant or RevokeGrant operation.

cgrsGrantToken :: Lens' CreateGrantResponse (Maybe Text)

The grant token. For more information, see Grant Tokens in the AWS Key Management Service Developer Guide .

cgrsResponseStatus :: Lens' CreateGrantResponse Int

- | The response status code.