Network.AWS.KMS.CreateKey

Description

Creates a customer master key (CMK).

You can use a CMK to encrypt small amounts of data (4 KiB or less) directly, but CMKs are more commonly used to encrypt data encryption keys (DEKs), which are used to encrypt raw data. For more information about DEKs and the difference between CMKs and DEKs, see the following:

The GenerateDataKey operation

AWS Key Management Service Concepts in the AWS Key Management Service Developer Guide

Creating a Request

createKey :: CreateKey

Creates a value of CreateKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ckOrigin - The source of the CMK's key material. The default is AWS_KMS , which means AWS KMS creates the key material. When this parameter is set to EXTERNAL , the request creates a CMK without key material so that you can import key material from your existing key management infrastructure. For more information about importing key material into AWS KMS, see Importing Key Material in the AWS Key Management Service Developer Guide . The CMK's Origin is immutable and is set when the CMK is created.

ckKeyUsage - The intended use of the CMK. You can use CMKs only for symmetric encryption and decryption.

ckBypassPolicyLockoutSafetyCheck - A flag to indicate whether to bypass the key policy lockout safety check. Important: Setting this value to true increases the likelihood that the CMK becomes unmanageable. Do not set this value to true indiscriminately. For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide . Use this parameter only when you include a policy in the request and you intend to prevent the principal that is making the request from making a subsequent PutKeyPolicy request on the CMK. The default value is false.

ckPolicy - The key policy to attach to the CMK. If you specify a policy and do not set BypassPolicyLockoutSafetyCheck to true, the policy must meet the following criteria: * It must allow the principal that is making the CreateKey request to make a subsequent PutKeyPolicy request on the CMK. This reduces the likelihood that the CMK becomes unmanageable. For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide . * The principals that are specified in the key policy must exist and be visible to AWS KMS. When you create a new AWS principal (for example, an IAM user or role), you might need to enforce a delay before specifying the new principal in a key policy because the new principal might not immediately be visible to AWS KMS. For more information, see Changes that I make are not always immediately visible in the IAM User Guide . If you do not specify a policy, AWS KMS attaches a default key policy to the CMK. For more information, see Default Key Policy in the AWS Key Management Service Developer Guide . The policy size limit is 32 KiB (32768 bytes).

ckDescription - A description of the CMK. Use a description that helps you decide whether the CMK is appropriate for a task.

ckTags - One or more tags. Each tag consists of a tag key and a tag value. Tag keys and tag values are both required, but tag values can be empty (null) strings. Use this parameter to tag the CMK when it is created. Alternately, you can omit this parameter and instead tag the CMK after it is created using TagResource .

data CreateKey

See: createKey smart constructor.

Request Lenses

ckOrigin :: Lens' CreateKey (Maybe OriginType)

The source of the CMK's key material. The default is AWS_KMS , which means AWS KMS creates the key material. When this parameter is set to EXTERNAL , the request creates a CMK without key material so that you can import key material from your existing key management infrastructure. For more information about importing key material into AWS KMS, see Importing Key Material in the AWS Key Management Service Developer Guide . The CMK's Origin is immutable and is set when the CMK is created.

ckKeyUsage :: Lens' CreateKey (Maybe KeyUsageType)

The intended use of the CMK. You can use CMKs only for symmetric encryption and decryption.

ckBypassPolicyLockoutSafetyCheck :: Lens' CreateKey (Maybe Bool)

A flag to indicate whether to bypass the key policy lockout safety check. Important: Setting this value to true increases the likelihood that the CMK becomes unmanageable. Do not set this value to true indiscriminately. For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide . Use this parameter only when you include a policy in the request and you intend to prevent the principal that is making the request from making a subsequent PutKeyPolicy request on the CMK. The default value is false.

ckPolicy :: Lens' CreateKey (Maybe Text)

The key policy to attach to the CMK. If you specify a policy and do not set BypassPolicyLockoutSafetyCheck to true, the policy must meet the following criteria: * It must allow the principal that is making the CreateKey request to make a subsequent PutKeyPolicy request on the CMK. This reduces the likelihood that the CMK becomes unmanageable. For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide . * The principals that are specified in the key policy must exist and be visible to AWS KMS. When you create a new AWS principal (for example, an IAM user or role), you might need to enforce a delay before specifying the new principal in a key policy because the new principal might not immediately be visible to AWS KMS. For more information, see Changes that I make are not always immediately visible in the IAM User Guide . If you do not specify a policy, AWS KMS attaches a default key policy to the CMK. For more information, see Default Key Policy in the AWS Key Management Service Developer Guide . The policy size limit is 32 KiB (32768 bytes).

ckDescription :: Lens' CreateKey (Maybe Text)

A description of the CMK. Use a description that helps you decide whether the CMK is appropriate for a task.

ckTags :: Lens' CreateKey [Tag]

One or more tags. Each tag consists of a tag key and a tag value. Tag keys and tag values are both required, but tag values can be empty (null) strings. Use this parameter to tag the CMK when it is created. Alternately, you can omit this parameter and instead tag the CMK after it is created using TagResource .

Destructuring the Response

createKeyResponse

Arguments

:: Int
ckrsResponseStatus

-> CreateKeyResponse

Creates a value of CreateKeyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ckrsKeyMetadata - Metadata associated with the CMK.

ckrsResponseStatus - -- | The response status code.

data CreateKeyResponse

See: createKeyResponse smart constructor.

Response Lenses

ckrsKeyMetadata :: Lens' CreateKeyResponse (Maybe KeyMetadata)

Metadata associated with the CMK.

ckrsResponseStatus :: Lens' CreateKeyResponse Int

- | The response status code.