Network.AWS.KMS.Encrypt

Description

Encrypts plaintext into ciphertext by using a customer master key. The Encrypt function has two primary use cases: You can encrypt up to 4 KB of arbitrary data such as an RSA key, a database password, or other sensitive customer information. If you are moving encrypted data from one region to another, you can use this API to encrypt in the new region the plaintext data key that was used to encrypt the data in the original region. This provides you with an encrypted copy of the data key that can be decrypted in the new region and used there to decrypt the encrypted data.

Unless you are moving encrypted data from one region to another, you don't use this function to encrypt a generated data key within a region. You retrieve data keys already encrypted by calling the GenerateDataKey or GenerateDataKeyWithoutPlaintext function. Data keys don't need to be encrypted again by calling Encrypt.

If you want to encrypt data locally in your application, you can use the GenerateDataKey function to return a plaintext data encryption key and a copy of the key encrypted under the customer master key (CMK) of your choosing.

http://docs.aws.amazon.com/kms/latest/APIReference/API_Encrypt.html

Request

data Encrypt

Request constructor

encrypt

Arguments

:: Text
eKeyId

-> Base64
ePlaintext

-> Encrypt

Encrypt constructor.

The fields accessible through corresponding lenses are:

eEncryptionContext :: HashMap Text Text

eGrantTokens :: [Text]

eKeyId :: Text

ePlaintext :: Base64

Request lenses

eEncryptionContext :: Lens' Encrypt (HashMap Text Text)

Name/value pair that specifies the encryption context to be used for authenticated encryption. If used here, the same value must be supplied to the Decrypt API or decryption will fail. For more information, see EncryptionContext.

eGrantTokens :: Lens' Encrypt [Text]

For more information, see Grant Tokens.

eKeyId :: Lens' Encrypt Text

A unique identifier for the customer master key. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/". Key ARN Example - arn:aws:kms:us-east-1:123456789012:key12345678-1234-1234-1234-123456789012 Alias ARN Example - arn:aws:kms:us-east-1:123456789012:alias/MyAliasName Globally Unique Key ID Example - 12345678-1234-1234-123456789012 Alias Name Example - alias/MyAliasName

ePlaintext :: Lens' Encrypt Base64

Data to be encrypted.

Response

data EncryptResponse

Response constructor

encryptResponse :: EncryptResponse

EncryptResponse constructor.

The fields accessible through corresponding lenses are:

erCiphertextBlob :: Maybe Base64

erKeyId :: Maybe Text

Response lenses

erCiphertextBlob :: Lens' EncryptResponse (Maybe Base64)

The encrypted plaintext. If you are using the CLI, the value is Base64 encoded. Otherwise, it is not encoded.

erKeyId :: Lens' EncryptResponse (Maybe Text)

The ID of the key used during encryption.