Network.AWS.KMS.GenerateDataKey

Description

Returns a data encryption key that you can use in your application to encrypt data locally.

You must specify the customer master key (CMK) under which to generate the data key. You must also specify the length of the data key using either the KeySpec or NumberOfBytes field. You must specify one field or the other, but not both. For common key lengths (128-bit and 256-bit symmetric keys), we recommend that you use KeySpec .

This operation returns a plaintext copy of the data key in the Plaintext field of the response, and an encrypted copy of the data key in the CiphertextBlob field. The data key is encrypted under the CMK specified in the KeyId field of the request.

We recommend that you use the following pattern to encrypt data locally in your application:

Use this operation (GenerateDataKey ) to retrieve a data encryption key.

Use the plaintext data encryption key (returned in the Plaintext field of the response) to encrypt data locally, then erase the plaintext data key from memory.

Store the encrypted data key (returned in the CiphertextBlob field of the response) alongside the locally encrypted data.

To decrypt data locally:

Use the Decrypt operation to decrypt the encrypted data key into a plaintext copy of the data key.

Use the plaintext data key to decrypt data locally, then erase the plaintext data key from memory.

To return only an encrypted copy of the data key, use GenerateDataKeyWithoutPlaintext . To return a random byte string that is cryptographically secure, use GenerateRandom .

If you use the optional EncryptionContext field, you must store at least enough information to be able to reconstruct the full encryption context when you later send the ciphertext to the Decrypt operation. It is a good practice to choose an encryption context that you can reconstruct on the fly to better secure the ciphertext. For more information, see Encryption Context in the AWS Key Management Service Developer Guide .

Creating a Request

generateDataKey

Arguments

:: Text
gdkKeyId

-> GenerateDataKey

Creates a value of GenerateDataKey with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdkKeySpec - The length of the data encryption key. Use AES_128 to generate a 128-bit symmetric key, or AES_256 to generate a 256-bit symmetric key.

gdkEncryptionContext - A set of key-value pairs that represents additional authenticated data. For more information, see Encryption Context in the AWS Key Management Service Developer Guide .

gdkNumberOfBytes - The length of the data encryption key in bytes. For example, use the value 64 to generate a 512-bit data key (64 bytes is 512 bits). For common key lengths (128-bit and 256-bit symmetric keys), we recommend that you use the KeySpec field instead of this one.

gdkGrantTokens - A list of grant tokens. For more information, see Grant Tokens in the AWS Key Management Service Developer Guide .

gdkKeyId - The identifier of the CMK under which to generate and encrypt the data encryption key. A valid identifier is the unique key ID or the Amazon Resource Name (ARN) of the CMK, or the alias name or ARN of an alias that refers to the CMK. Examples: * Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab * CMK ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab * Alias name: alias/ExampleAlias * Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

data GenerateDataKey

See: generateDataKey smart constructor.

Request Lenses

gdkKeySpec :: Lens' GenerateDataKey (Maybe DataKeySpec)

The length of the data encryption key. Use AES_128 to generate a 128-bit symmetric key, or AES_256 to generate a 256-bit symmetric key.

gdkEncryptionContext :: Lens' GenerateDataKey (HashMap Text Text)

A set of key-value pairs that represents additional authenticated data. For more information, see Encryption Context in the AWS Key Management Service Developer Guide .

gdkNumberOfBytes :: Lens' GenerateDataKey (Maybe Natural)

The length of the data encryption key in bytes. For example, use the value 64 to generate a 512-bit data key (64 bytes is 512 bits). For common key lengths (128-bit and 256-bit symmetric keys), we recommend that you use the KeySpec field instead of this one.

gdkGrantTokens :: Lens' GenerateDataKey [Text]

A list of grant tokens. For more information, see Grant Tokens in the AWS Key Management Service Developer Guide .

gdkKeyId :: Lens' GenerateDataKey Text

The identifier of the CMK under which to generate and encrypt the data encryption key. A valid identifier is the unique key ID or the Amazon Resource Name (ARN) of the CMK, or the alias name or ARN of an alias that refers to the CMK. Examples: * Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab * CMK ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab * Alias name: alias/ExampleAlias * Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

Destructuring the Response

generateDataKeyResponse

Arguments

:: Int
gdkrsResponseStatus

-> Text
gdkrsKeyId

-> ByteString
gdkrsPlaintext

-> ByteString
gdkrsCiphertextBlob

-> GenerateDataKeyResponse

Creates a value of GenerateDataKeyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdkrsResponseStatus - -- | The response status code.

gdkrsKeyId - The identifier of the CMK under which the data encryption key was generated and encrypted.

gdkrsPlaintext - The data encryption key. Use this data key for local encryption and decryption, then remove it from memory as soon as possible.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

gdkrsCiphertextBlob - The encrypted data encryption key.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

data GenerateDataKeyResponse

See: generateDataKeyResponse smart constructor.

Response Lenses

gdkrsResponseStatus :: Lens' GenerateDataKeyResponse Int

- | The response status code.

gdkrsKeyId :: Lens' GenerateDataKeyResponse Text

The identifier of the CMK under which the data encryption key was generated and encrypted.

gdkrsPlaintext :: Lens' GenerateDataKeyResponse ByteString

The data encryption key. Use this data key for local encryption and decryption, then remove it from memory as soon as possible.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

gdkrsCiphertextBlob :: Lens' GenerateDataKeyResponse ByteString

The encrypted data encryption key.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.