Network.AWS.KMS.ListAliases

Description

Lists all of the key aliases in the account.

This operation returns paginated results.

Creating a Request

listAliases :: ListAliases

Creates a value of ListAliases with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laMarker - Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

laLimit - Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer. This value is optional. If you include a value, it must be between 1 and 100, inclusive. If you do not include a value, it defaults to 50.

data ListAliases

See: listAliases smart constructor.

Request Lenses

laMarker :: Lens' ListAliases (Maybe Text)

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

laLimit :: Lens' ListAliases (Maybe Natural)

Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer. This value is optional. If you include a value, it must be between 1 and 100, inclusive. If you do not include a value, it defaults to 50.

Destructuring the Response

listAliasesResponse

Arguments

:: Int
larsResponseStatus

-> ListAliasesResponse

Creates a value of ListAliasesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

larsTruncated - A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To retrieve more items, pass the value of the NextMarker element in this response to the Marker parameter in a subsequent request.

larsAliases - A list of key aliases in the user's account.

larsNextMarker - When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

larsResponseStatus - -- | The response status code.

data ListAliasesResponse

See: listAliasesResponse smart constructor.

Response Lenses

larsTruncated :: Lens' ListAliasesResponse (Maybe Bool)

A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To retrieve more items, pass the value of the NextMarker element in this response to the Marker parameter in a subsequent request.

larsAliases :: Lens' ListAliasesResponse [AliasListEntry]

A list of key aliases in the user's account.

larsNextMarker :: Lens' ListAliasesResponse (Maybe Text)

When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

larsResponseStatus :: Lens' ListAliasesResponse Int

- | The response status code.