Network.AWS.KMS.ListGrants

Description

List the grants for a specified key.

This operation returns paginated results.

Creating a Request

listGrants

Arguments

:: Text
lgKeyId

-> ListGrants

Creates a value of ListGrants with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lgMarker - Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

lgLimit - Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer. This value is optional. If you include a value, it must be between 1 and 100, inclusive. If you do not include a value, it defaults to 50.

lgKeyId - A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key. * Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012 * Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

data ListGrants

See: listGrants smart constructor.

Request Lenses

lgMarker :: Lens' ListGrants (Maybe Text)

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

lgLimit :: Lens' ListGrants (Maybe Natural)

Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer. This value is optional. If you include a value, it must be between 1 and 100, inclusive. If you do not include a value, it defaults to 50.

lgKeyId :: Lens' ListGrants Text

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key. * Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012 * Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Destructuring the Response

listGrantsResponse :: ListGrantsResponse

Creates a value of ListGrantsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lgTruncated - A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To retrieve more items, pass the value of the NextMarker element in this response to the Marker parameter in a subsequent request.

lgGrants - A list of grants.

lgNextMarker - When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

data ListGrantsResponse

See: listGrantsResponse smart constructor.

Response Lenses

lgTruncated :: Lens' ListGrantsResponse (Maybe Bool)

A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To retrieve more items, pass the value of the NextMarker element in this response to the Marker parameter in a subsequent request.

lgGrants :: Lens' ListGrantsResponse [GrantListEntry]

A list of grants.

lgNextMarker :: Lens' ListGrantsResponse (Maybe Text)

When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.