Network.AWS.KMS.ListKeyPolicies

Description

Retrieves a list of policies attached to a key.

This operation returns paginated results.

Creating a Request

listKeyPolicies

Arguments

:: Text
lkpKeyId

-> ListKeyPolicies

Creates a value of ListKeyPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lkpMarker - Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

lkpLimit - Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer. This value is optional. If you include a value, it must be between 1 and 1000, inclusive. If you do not include a value, it defaults to 100. Currently only 1 policy can be attached to a key.

lkpKeyId - A unique identifier for the customer master key (CMK). You can use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples: * Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab * Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

data ListKeyPolicies

See: listKeyPolicies smart constructor.

Request Lenses

lkpMarker :: Lens' ListKeyPolicies (Maybe Text)

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

lkpLimit :: Lens' ListKeyPolicies (Maybe Natural)

Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer. This value is optional. If you include a value, it must be between 1 and 1000, inclusive. If you do not include a value, it defaults to 100. Currently only 1 policy can be attached to a key.

lkpKeyId :: Lens' ListKeyPolicies Text

A unique identifier for the customer master key (CMK). You can use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples: * Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab * Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Destructuring the Response

listKeyPoliciesResponse

Arguments

:: Int
lkprsResponseStatus

-> ListKeyPoliciesResponse

Creates a value of ListKeyPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lkprsPolicyNames - A list of policy names. Currently, there is only one policy and it is named Default.

lkprsTruncated - A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To retrieve more items, pass the value of the NextMarker element in this response to the Marker parameter in a subsequent request.

lkprsNextMarker - When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

lkprsResponseStatus - -- | The response status code.

data ListKeyPoliciesResponse

See: listKeyPoliciesResponse smart constructor.

Response Lenses

lkprsPolicyNames :: Lens' ListKeyPoliciesResponse [Text]

A list of policy names. Currently, there is only one policy and it is named Default.

lkprsTruncated :: Lens' ListKeyPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To retrieve more items, pass the value of the NextMarker element in this response to the Marker parameter in a subsequent request.

lkprsNextMarker :: Lens' ListKeyPoliciesResponse (Maybe Text)

When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

lkprsResponseStatus :: Lens' ListKeyPoliciesResponse Int

- | The response status code.