Network.AWS.KMS.ReEncrypt

Description

Encrypts data on the server side with a new customer master key (CMK) without exposing the plaintext of the data on the client side. The data is first decrypted and then reencrypted. You can also use this operation to change the encryption context of a ciphertext.

Unlike other operations, ReEncrypt is authorized twice, once as ReEncryptFrom on the source CMK and once as ReEncryptTo on the destination CMK. We recommend that you include the "kms:ReEncrypt*" permission in your key policies to permit reencryption from or to the CMK. This permission is automatically included in the key policy when you create a CMK through the console, but you must include it manually when you create a CMK programmatically or when you set a key policy with the PutKeyPolicy operation.

Creating a Request

reEncrypt

Arguments

:: ByteString
reCiphertextBlob

-> Text
reDestinationKeyId

-> ReEncrypt

Creates a value of ReEncrypt with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

reDestinationEncryptionContext - Encryption context to use when the data is reencrypted.

reSourceEncryptionContext - Encryption context used to encrypt and decrypt the data specified in the CiphertextBlob parameter.

reGrantTokens - A list of grant tokens. For more information, see Grant Tokens in the AWS Key Management Service Developer Guide .

reCiphertextBlob - Ciphertext of the data to reencrypt.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

reDestinationKeyId - A unique identifier for the CMK to use to reencrypt the data. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias". * Key ARN Example - arn:aws:kms:us-east-1:123456789012:key12345678-1234-1234-1234-123456789012 * Alias ARN Example - arn:aws:kms:us-east-1:123456789012:aliasMyAliasName * Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012 * Alias Name Example - aliasMyAliasName

data ReEncrypt

See: reEncrypt smart constructor.

Request Lenses

reDestinationEncryptionContext :: Lens' ReEncrypt (HashMap Text Text)

Encryption context to use when the data is reencrypted.

reSourceEncryptionContext :: Lens' ReEncrypt (HashMap Text Text)

Encryption context used to encrypt and decrypt the data specified in the CiphertextBlob parameter.

reGrantTokens :: Lens' ReEncrypt [Text]

A list of grant tokens. For more information, see Grant Tokens in the AWS Key Management Service Developer Guide .

reCiphertextBlob :: Lens' ReEncrypt ByteString

Ciphertext of the data to reencrypt.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

reDestinationKeyId :: Lens' ReEncrypt Text

A unique identifier for the CMK to use to reencrypt the data. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias". * Key ARN Example - arn:aws:kms:us-east-1:123456789012:key12345678-1234-1234-1234-123456789012 * Alias ARN Example - arn:aws:kms:us-east-1:123456789012:aliasMyAliasName * Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012 * Alias Name Example - aliasMyAliasName

Destructuring the Response

reEncryptResponse

Arguments

:: Int
rersResponseStatus

-> ReEncryptResponse

Creates a value of ReEncryptResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rersSourceKeyId - Unique identifier of the CMK used to originally encrypt the data.

rersKeyId - Unique identifier of the CMK used to reencrypt the data.

rersCiphertextBlob - The reencrypted data.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

rersResponseStatus - -- | The response status code.

data ReEncryptResponse

See: reEncryptResponse smart constructor.

Response Lenses

rersSourceKeyId :: Lens' ReEncryptResponse (Maybe Text)

Unique identifier of the CMK used to originally encrypt the data.

rersKeyId :: Lens' ReEncryptResponse (Maybe Text)

Unique identifier of the CMK used to reencrypt the data.

rersCiphertextBlob :: Lens' ReEncryptResponse (Maybe ByteString)

The reencrypted data.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

rersResponseStatus :: Lens' ReEncryptResponse Int

- | The response status code.