Network.AWS.KMS.Types

Description

Service Configuration

kms :: Service

API version 2014-11-01 of the Amazon Key Management Service SDK configuration.

Errors

_InvalidMarkerException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the marker that specifies where pagination should next begin is not valid.

_KMSInvalidStateException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the state of the specified resource is not valid for this request.

For more information about how key state affects the use of a CMK, see How Key State Affects Use of a Customer Master Key in the AWS Key Management Service Developer Guide .

_InvalidKeyUsageException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified KeySpec value is not valid.

_MalformedPolicyDocumentException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified policy is not syntactically or semantically correct.

_UnsupportedOperationException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because a specified parameter is not supported or a specified resource is not valid for this operation.

_DisabledException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified CMK is not enabled.

_KeyUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified CMK was not available. The request can be retried.

_IncorrectKeyMaterialException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the provided key material is invalid or is not the same key material that was previously imported into this customer master key (CMK).

_KMSInternalException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because an internal exception occurred. The request can be retried.

_TagException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because one or more tags are not valid.

_InvalidImportTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the provided import token is invalid or is associated with a different customer master key (CMK).

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified entity or resource could not be found.

_InvalidAliasNameException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified alias name is not valid.

_InvalidGrantIdException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified GrantId is not valid.

_InvalidGrantTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified grant token is not valid.

_InvalidARNException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because a specified ARN was not valid.

_DependencyTimeoutException :: AsError a => Getting (First ServiceError) a ServiceError

The system timed out while trying to fulfill the request. The request can be retried.

_ExpiredImportTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the provided import token is expired. Use GetParametersForImport to retrieve a new import token and public key, use the new public key to encrypt the key material, and then try the request again.

_InvalidCiphertextException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because the specified ciphertext has been corrupted or is otherwise invalid.

_AlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it attempted to create a resource that already exists.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because a limit was exceeded. For more information, see Limits in the AWS Key Management Service Developer Guide .

AlgorithmSpec

data AlgorithmSpec

Constructors

RsaesOaepSha1

RsaesOaepSha256

RsaesPKCS1V15

DataKeySpec

data DataKeySpec

Constructors

AES128

AES256

ExpirationModelType

data ExpirationModelType

Constructors

KeyMaterialDoesNotExpire

KeyMaterialExpires

GrantOperation

data GrantOperation

Constructors

CreateGrant

Decrypt

DescribeKey

Encrypt

GenerateDataKey

GenerateDataKeyWithoutPlaintext

ReEncryptFrom

ReEncryptTo

RetireGrant

KeyManagerType

data KeyManagerType

Constructors

AWS

Customer

KeyState

data KeyState

Constructors

Disabled

Enabled

PendingDeletion

PendingImport

KeyUsageType

data KeyUsageType

Constructors

EncryptDecrypt

OriginType

data OriginType

Constructors

AWSKMS

External

WrappingKeySpec

data WrappingKeySpec

Constructors

Rsa2048

AliasListEntry

data AliasListEntry

Contains information about an alias.

See: aliasListEntry smart constructor.

aliasListEntry :: AliasListEntry

Creates a value of AliasListEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aleTargetKeyId - String that contains the key identifier referred to by the alias.

aleAliasName - String that contains the alias.

aleAliasARN - String that contains the key ARN.

aleTargetKeyId :: Lens' AliasListEntry (Maybe Text)

String that contains the key identifier referred to by the alias.

aleAliasName :: Lens' AliasListEntry (Maybe Text)

String that contains the alias.

aleAliasARN :: Lens' AliasListEntry (Maybe Text)

String that contains the key ARN.

GrantConstraints

data GrantConstraints

A structure that you can use to allow certain operations in the grant only when the desired encryption context is present. For more information about encryption context, see Encryption Context in the AWS Key Management Service Developer Guide .

Grant constraints apply only to operations that accept encryption context as input. For example, the DescribeKey operation does not accept encryption context as input. A grant that allows the DescribeKey operation does so regardless of the grant constraints. In constrast, the Encrypt operation accepts encryption context as input. A grant that allows the Encrypt operation does so only when the encryption context of the Encrypt operation satisfies the grant constraints.

See: grantConstraints smart constructor.

grantConstraints :: GrantConstraints

Creates a value of GrantConstraints with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcEncryptionContextEquals - A list of key-value pairs that must be present in the encryption context of certain subsequent operations that the grant allows. When certain subsequent operations allowed by the grant include encryption context that matches this list, the grant allows the operation. Otherwise, the grant does not allow the operation.

gcEncryptionContextSubset - A list of key-value pairs, all of which must be present in the encryption context of certain subsequent operations that the grant allows. When certain subsequent operations allowed by the grant include encryption context that matches this list or is a superset of this list, the grant allows the operation. Otherwise, the grant does not allow the operation.

gcEncryptionContextEquals :: Lens' GrantConstraints (HashMap Text Text)

A list of key-value pairs that must be present in the encryption context of certain subsequent operations that the grant allows. When certain subsequent operations allowed by the grant include encryption context that matches this list, the grant allows the operation. Otherwise, the grant does not allow the operation.

gcEncryptionContextSubset :: Lens' GrantConstraints (HashMap Text Text)

A list of key-value pairs, all of which must be present in the encryption context of certain subsequent operations that the grant allows. When certain subsequent operations allowed by the grant include encryption context that matches this list or is a superset of this list, the grant allows the operation. Otherwise, the grant does not allow the operation.

GrantListEntry

data GrantListEntry

Contains information about an entry in a list of grants.

See: grantListEntry smart constructor.

grantListEntry :: GrantListEntry

Creates a value of GrantListEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gleKeyId - The unique identifier for the customer master key (CMK) to which the grant applies.

gleRetiringPrincipal - The principal that can retire the grant.

gleIssuingAccount - The AWS account under which the grant was issued.

gleGrantId - The unique identifier for the grant.

gleConstraints - A list of key-value pairs that must be present in the encryption context of certain subsequent operations that the grant allows.

gleGranteePrincipal - The principal that receives the grant's permissions.

gleName - The friendly name that identifies the grant. If a name was provided in the CreateGrant request, that name is returned. Otherwise this value is null.

gleCreationDate - The date and time when the grant was created.

gleOperations - The list of operations permitted by the grant.

gleKeyId :: Lens' GrantListEntry (Maybe Text)

The unique identifier for the customer master key (CMK) to which the grant applies.

gleRetiringPrincipal :: Lens' GrantListEntry (Maybe Text)

The principal that can retire the grant.

gleIssuingAccount :: Lens' GrantListEntry (Maybe Text)

The AWS account under which the grant was issued.

gleGrantId :: Lens' GrantListEntry (Maybe Text)

The unique identifier for the grant.

gleConstraints :: Lens' GrantListEntry (Maybe GrantConstraints)

A list of key-value pairs that must be present in the encryption context of certain subsequent operations that the grant allows.

gleGranteePrincipal :: Lens' GrantListEntry (Maybe Text)

The principal that receives the grant's permissions.

gleName :: Lens' GrantListEntry (Maybe Text)

The friendly name that identifies the grant. If a name was provided in the CreateGrant request, that name is returned. Otherwise this value is null.

gleCreationDate :: Lens' GrantListEntry (Maybe UTCTime)

The date and time when the grant was created.

gleOperations :: Lens' GrantListEntry [GrantOperation]

The list of operations permitted by the grant.

KeyListEntry

data KeyListEntry

Contains information about each entry in the key list.

See: keyListEntry smart constructor.

keyListEntry :: KeyListEntry

Creates a value of KeyListEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kleKeyId - Unique identifier of the key.

kleKeyARN - ARN of the key.

kleKeyId :: Lens' KeyListEntry (Maybe Text)

Unique identifier of the key.

kleKeyARN :: Lens' KeyListEntry (Maybe Text)

ARN of the key.

KeyMetadata

data KeyMetadata

Contains metadata about a customer master key (CMK).

This data type is used as a response element for the CreateKey and DescribeKey operations.

See: keyMetadata smart constructor.

keyMetadata

Arguments

:: Text
kmKeyId

-> KeyMetadata

Creates a value of KeyMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

kmOrigin - The source of the CMK's key material. When this value is AWS_KMS , AWS KMS created the key material. When this value is EXTERNAL , the key material was imported from your existing key management infrastructure or the CMK lacks key material.

kmExpirationModel - Specifies whether the CMK's key material expires. This value is present only when Origin is EXTERNAL , otherwise this value is omitted.

kmKeyManager - The CMK's manager. CMKs are either customer-managed or AWS-managed. For more information about the difference, see Customer Master Keys in the AWS Key Management Service Developer Guide .

kmEnabled - Specifies whether the CMK is enabled. When KeyState is Enabled this value is true, otherwise it is false.

kmValidTo - The time at which the imported key material expires. When the key material expires, AWS KMS deletes the key material and the CMK becomes unusable. This value is present only for CMKs whose Origin is EXTERNAL and whose ExpirationModel is KEY_MATERIAL_EXPIRES , otherwise this value is omitted.

kmARN - The Amazon Resource Name (ARN) of the CMK. For examples, see AWS Key Management Service (AWS KMS) in the Example ARNs section of the AWS General Reference .

kmKeyState - The state of the CMK. For more information about how key state affects the use of a CMK, see How Key State Affects the Use of a Customer Master Key in the AWS Key Management Service Developer Guide .

kmAWSAccountId - The twelve-digit account ID of the AWS account that owns the CMK.

kmKeyUsage - The cryptographic operations for which you can use the CMK. Currently the only allowed value is ENCRYPT_DECRYPT , which means you can use the CMK for the Encrypt and Decrypt operations.

kmCreationDate - The date and time when the CMK was created.

kmDeletionDate - The date and time after which AWS KMS deletes the CMK. This value is present only when KeyState is PendingDeletion , otherwise this value is omitted.

kmDescription - The description of the CMK.

kmKeyId - The globally unique identifier for the CMK.

kmOrigin :: Lens' KeyMetadata (Maybe OriginType)

The source of the CMK's key material. When this value is AWS_KMS , AWS KMS created the key material. When this value is EXTERNAL , the key material was imported from your existing key management infrastructure or the CMK lacks key material.

kmExpirationModel :: Lens' KeyMetadata (Maybe ExpirationModelType)

Specifies whether the CMK's key material expires. This value is present only when Origin is EXTERNAL , otherwise this value is omitted.

kmKeyManager :: Lens' KeyMetadata (Maybe KeyManagerType)

The CMK's manager. CMKs are either customer-managed or AWS-managed. For more information about the difference, see Customer Master Keys in the AWS Key Management Service Developer Guide .

kmEnabled :: Lens' KeyMetadata (Maybe Bool)

Specifies whether the CMK is enabled. When KeyState is Enabled this value is true, otherwise it is false.

kmValidTo :: Lens' KeyMetadata (Maybe UTCTime)

The time at which the imported key material expires. When the key material expires, AWS KMS deletes the key material and the CMK becomes unusable. This value is present only for CMKs whose Origin is EXTERNAL and whose ExpirationModel is KEY_MATERIAL_EXPIRES , otherwise this value is omitted.

kmARN :: Lens' KeyMetadata (Maybe Text)

The Amazon Resource Name (ARN) of the CMK. For examples, see AWS Key Management Service (AWS KMS) in the Example ARNs section of the AWS General Reference .

kmKeyState :: Lens' KeyMetadata (Maybe KeyState)

The state of the CMK. For more information about how key state affects the use of a CMK, see How Key State Affects the Use of a Customer Master Key in the AWS Key Management Service Developer Guide .

kmAWSAccountId :: Lens' KeyMetadata (Maybe Text)

The twelve-digit account ID of the AWS account that owns the CMK.

kmKeyUsage :: Lens' KeyMetadata (Maybe KeyUsageType)

The cryptographic operations for which you can use the CMK. Currently the only allowed value is ENCRYPT_DECRYPT , which means you can use the CMK for the Encrypt and Decrypt operations.

kmCreationDate :: Lens' KeyMetadata (Maybe UTCTime)

The date and time when the CMK was created.

kmDeletionDate :: Lens' KeyMetadata (Maybe UTCTime)

The date and time after which AWS KMS deletes the CMK. This value is present only when KeyState is PendingDeletion , otherwise this value is omitted.

kmDescription :: Lens' KeyMetadata (Maybe Text)

The description of the CMK.

kmKeyId :: Lens' KeyMetadata Text

The globally unique identifier for the CMK.

ListGrantsResponse

data ListGrantsResponse

See: listGrantsResponse smart constructor.

listGrantsResponse :: ListGrantsResponse

Creates a value of ListGrantsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lgTruncated - A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To retrieve more items, pass the value of the NextMarker element in this response to the Marker parameter in a subsequent request.

lgGrants - A list of grants.

lgNextMarker - When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

lgTruncated :: Lens' ListGrantsResponse (Maybe Bool)

A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To retrieve more items, pass the value of the NextMarker element in this response to the Marker parameter in a subsequent request.

lgGrants :: Lens' ListGrantsResponse [GrantListEntry]

A list of grants.

lgNextMarker :: Lens' ListGrantsResponse (Maybe Text)

When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

Tag

data Tag

A key-value pair. A tag consists of a tag key and a tag value. Tag keys and tag values are both required, but tag values can be empty (null) strings.

See: tag smart constructor.

tag

Arguments

:: Text
tagTagKey

-> Text
tagTagValue

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagTagKey - The key of the tag.

tagTagValue - The value of the tag.

tagTagKey :: Lens' Tag Text

The key of the tag.

tagTagValue :: Lens' Tag Text

The value of the tag.